✨: add duplicate budget functionality.
- code slitting into multi routes.
- some cleanups
