+++ src/app/app/expenses/page.tsx
@@ -1,4 +1,14 @@
-export default function ExpensesPage() {
+import { connection } from "next/server";
+import { api } from "@/trpc/server";
+
+export default async function ExpensesPage() {
+  await connection();
+
+  void api.expense.getAll.prefetch({
+    page: 1,
+    pageSize: 100,
+  });
+
   return (
     <>
       <div className="grid auto-rows-min gap-4 md:grid-cols-3">
