+++ src/server/api/routers/budget.ts
@@ -20,27 +20,31 @@ export const budgetRouter = createTRPCRouter({
       const { page, pageSize } = input;
       const offset = (page - 1) * pageSize;
 
-      // Get total count
-      const countResult = await ctx.db
-        .select({ count: sql<number>`count(*)` })
-        .from(budget)
-        .where(eq(budget.user_id, ctx.user.id));
-      const count = countResult[0]?.count ?? 0;
-
-      // Get paginated results
-      const items = await ctx.db
-        .select()
+      const results = await ctx.db
+        .select({
+          id: budget.id,
+          name: budget.name,
+          description: budget.description,
+          amount: budget.amount,
+          user_id: budget.user_id,
+          created_at: budget.created_at,
+          updated_at: budget.updated_at,
+          totalCount: sql<number>`count(*) over()`,
+        })
         .from(budget)
         .where(eq(budget.user_id, ctx.user.id))
         .orderBy(desc(budget.created_at))
         .limit(pageSize)
         .offset(offset);
 
+      const items = results.map(({ totalCount, ...item }) => item);
+      const totalItems = Number(results[0]?.totalCount ?? 0);
+
       return {
         items,
-        totalPages: Math.ceil(Number(count) / pageSize),
+        totalPages: Math.ceil(totalItems / pageSize),
         currentPage: page,
-        totalItems: Number(count),
+        totalItems,
       };
     }),
 
@@ -87,6 +91,37 @@ export const budgetRouter = createTRPCRouter({
 
       return {
         success: result.length > 0,
+      };
+    }),
+
+  duplicate: privateProcedure
+    .input(z.object({ id: z.string() }))
+    .mutation(async ({ ctx, input }) => {
+      const [existingBudget] = await ctx.db
+        .select()
+        .from(budget)
+        .where(and(eq(budget.id, input.id), eq(budget.user_id, ctx.user.id)));
+
+      if (!existingBudget) {
+        return { success: false };
+      }
+
+      const [newBudget] = await ctx.db
+        .insert(budget)
+        .values({
+          user_id: ctx.user.id,
+          name: `${existingBudget.name} (Copy)`,
+          description: `${existingBudget.description} (Copy)`,
+          amount: existingBudget.amount,
+
+          created_at: new Date(),
+          updated_at: new Date(),
+        })
+        .returning();
+
+      return {
+        success: true,
+        budget: newBudget,
       };
     }),
 });
