+++ src/styles/globals.css
@@ -18,7 +18,7 @@
   --popover-foreground: oklch(0.1448 0 0);
   --primary: oklch(0.3192 0.0513 170.88); /* #113b2f */
   --primary-foreground: oklch(0.9851 0 0);
-  --secondary: oklch(0.815 0.1628 83.04); /* #f5b71b */
+  --secondary: oklch(0.5418 0.1036 167.9589); /* #113b2f */
   --secondary-foreground: oklch(0.2046 0 0);
   --muted: oklch(0.9702 0 0);
   --muted-foreground: oklch(0.5486 0 0);
